requirements.txt
aiofile
asyncpg
bcrypt
fastapi
passlib[bcrypt]
psycopg2-binary
pyotp
python-jose[cryptography]
pyyaml
SQLAlchemy
uvicorn
